Changes necessary to solve games on 64-bit architectures, as well as some compiler warnings fixed. -JJ
git-svn-id: https://gamescrafters.svn.sourceforge.net/svnroot/gamescrafters/GamesmanClassic/trunk@884 a668be5d-ef11-cb4c-8bfd-e08cbb24f2ba

+++ gamesman/src/gamesman.h
@@ -27,7 +27,7 @@
 #define  TRUE           1
 #endif
 
-typedef int POSITION;       /* This used to be determined by mTempTypedef.h */
+typedef long POSITION;      /* This used to be determined by mTempTypedef.h */
 typedef int MOVE;           /* but I decided to hardcode it for now, so as  */
 typedef int MOVES;          /* to make it a leeeetle bit easier and cleaner */
